+++ src/AttributedDI/AttributedDiServiceCollectionExtensions.cs
@@ -1,8 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
 
 namespace AttributedDI;
 
@@ -88,6 +86,17 @@ public static partial class AttributedDiServiceCollectionExtensions
     /// Adds AttributedDI services to the collection using the provided configuration.
     /// </summary>
     /// <param name="services">The service collection to configure.</param>
+    /// <returns>The service collection for chaining.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="services"/> is null.</exception>
+    public static IServiceCollection AddAttributedDi(this IServiceCollection services)
+    {
+        return AddAttributedDi(services, _ => { });
+    }
+
+    /// <summary>
+    /// Adds AttributedDI services to the collection using the provided configuration.
+    /// </summary>
+    /// <param name="services">The service collection to configure.</param>
     /// <param name="configure">A callback to configure AttributedDI options.</param>
     /// <returns>The service collection for chaining.</returns>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="services"/> or <paramref name="configure"/> is null.</exception>
@@ -132,60 +141,4 @@ public static partial class AttributedDiServiceCollectionExtensions
     }
 
     static partial void AddAttributedDiGeneratedModules(IServiceCollection services);
-}
-
-/// <summary>
-/// Represents configuration options for AttributedDI service registration.
-/// </summary>
-public class AttributedDiOptions
-{
-    private readonly HashSet<IServiceModule> modules = new(ModuleTypeComparer.Instance);
-
-    /// <summary>
-    /// Gets or sets whether deferred service support is enabled.
-    /// </summary>
-    public bool EnableDeferred { get; set; } = true;
-
-    /// <summary>
-    /// Gets or sets whether owned service support is enabled.
-    /// </summary>
-    public bool EnableOwned { get; set; } = true;
-
-    /// <summary>
-    /// Gets or sets whether factory delegate support is enabled.
-    /// </summary>
-    public bool EnableFactory { get; set; } = true;
-
-    /// <summary>
-    /// Gets or sets whether generated service modules should be registered.
-    /// </summary>
-    public bool RegisterGeneratedModules { get; set; } = true;
-
-    /// <summary>
-    /// Gets the additional modules to register.
-    /// </summary>
-    /// <remarks>Modules are deduplicated by concrete module type.</remarks>
-    public IReadOnlyCollection<IServiceModule> AdditionalModules => modules.ToArray();
-
-    /// <summary>
-    /// Adds an additional service module to register.
-    /// </summary>
-    /// <param name="module">The service module to add.</param>
-    /// <returns>The current options instance for chaining.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="module"/> is null.</exception>
-    public AttributedDiOptions AddAdditionalModules(IServiceModule module)
-    {
-        modules.Add(module);
-
-        return this;
-    }
-
-    private sealed class ModuleTypeComparer : IEqualityComparer<IServiceModule>
-    {
-        public static ModuleTypeComparer Instance { get; } = new();
-
-        public bool Equals(IServiceModule? x, IServiceModule? y) => x?.GetType() == y?.GetType();
-
-        public int GetHashCode([DisallowNull] IServiceModule obj) => HashCode.Combine(obj.GetType());
-    }
 }
